requirements.txt
pytesseract
numpy
Pillow
easyocr
flask
psycopg2-binary
flask-sqlalchemy
Flask-Migrate
python-dotenv
Flask-Script